Removes passing the clock and reset with o_clk and o_resetn done in a recent commit, updates README, adds tmp folder in order to support the scripted bash flow.
The reason we remove o_clk and o_resetn is that you usually would want to pass HCLK and HRESETn directly to the affected modules.
By passing the clock through the FreeAHB master, as done until now, it will make the critical path for the two signals longer than needed,
and you would add unneccessary clock propagation delay in the final implemented design. It also makes designs a bit neater.